Projet_Extracteur_de_Documents_Structurés_Nacim_Afrikou_Issam_Belhamadi/requirements.txt
pdfplumber
pydantic
python-dotenv
typer
rich
pytest
openai
streamlit
python-docx
openpyxl
pandas

